requirements.txt
# Core dependencies
together>=1.2.1
scikit-learn>=1.3.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
google-generativeai>=0.8.0
sentence-transformers>=2.2.2
google-genai

# Data processing
pandas>=2.0.0
numpy>=1.24.0
nltk>=3.8.1

# Visualization
streamlit>=1.28.0
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
click>=8.1.0

# Development
jupyter>=1.0.0
pytest>=7.4.0
black>=23.0.0
